Move secondary layout to a new component
