+++ public/sitemap-0.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://anavrin.app</loc><lastmod>2023-12-18T21:12:26.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://anavrin.app/blogs</loc><lastmod>2023-12-18T21:12:26.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://anavrin.app/dashboard</loc><lastmod>2023-12-18T21:12:26.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://anavrin.app/overview</loc><lastmod>2023-12-18T21:12:26.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://anavrin.app/portfolio</loc><lastmod>2023-12-18T21:12:26.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://anavrin.app/portfolios</loc><lastmod>2023-12-18T21:12:26.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://anavrin.app/register</loc><lastmod>2023-12-18T21:12:26.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://anavrin.app/simulator/cryptocurrency</loc><lastmod>2023-12-18T21:12:26.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://anavrin.app/simulator/dubai-financial-markets</loc><lastmod>2023-12-18T21:12:26.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://anavrin.app/user-settings</loc><lastmod>2023-12-18T21:12:26.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://anavrin.app/blogs/linea-airdrop-farming</loc><lastmod>2023-12-18T21:12:26.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://anavrin.app</loc><lastmod>2023-12-21T21:41:17.693Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://anavrin.app/blogs</loc><lastmod>2023-12-21T21:41:17.693Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://anavrin.app/dashboard</loc><lastmod>2023-12-21T21:41:17.693Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://anavrin.app/overview</loc><lastmod>2023-12-21T21:41:17.693Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://anavrin.app/portfolio</loc><lastmod>2023-12-21T21:41:17.693Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://anavrin.app/portfolios</loc><lastmod>2023-12-21T21:41:17.693Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://anavrin.app/register</loc><lastmod>2023-12-21T21:41:17.693Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://anavrin.app/simulator/cryptocurrency</loc><lastmod>2023-12-21T21:41:17.693Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://anavrin.app/simulator/dubai-financial-markets</loc><lastmod>2023-12-21T21:41:17.693Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://anavrin.app/user-settings</loc><lastmod>2023-12-21T21:41:17.693Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://anavrin.app/blogs/linea-airdrop-farming</loc><lastmod>2023-12-21T21:41:17.693Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
