+++ src/styles/index.css
@@ -105,10 +105,7 @@
 }
 
 .portfolio-secondary-panel {
-  @apply  bg-gray-200 p-2; 
-  /* @apply hidden sm:block border-gray-400; */
-  @apply border-gray-400;
-  @apply bg-charcoal-400 text-gray-300;
+  @apply bg-charcoal-400 text-gray-300 p-2;
   transition: width 0.5s ease;
 }
